requirements.txt
anyio==3.3.4
asgiref==3.4.1
attrs==21.2.0
bidict==0.21.4
click==8.0.3
cycler==0.10.0
fastapi==0.70.0
h11==0.12.0
idna==3.3
iniconfig==1.1.1
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
packaging==21.0
Pillow==8.4.0
pluggy==1.0.0
py==1.10.0
pydantic==1.8.2
pyparsing==3.0.2
pytest==6.2.5
pytest-asyncio==0.16.0
python-dateutil==2.8.2
python-engineio==4.3.0
python-multipart==0.0.5
python-socketio==5.4.1
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
sklearn==0.0
sniffio==1.2.0
starlette==0.16.0
threadpoolctl==3.0.0
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
websockets==10.0
